+++ GraphicsHomework/GraphicsHomework.h
@@ -30,7 +30,7 @@ public:
 	afx_msg void OnEditParams();
 	DECLARE_MESSAGE_MAP()
 public: // because I need them fron the view
-	float _a, _b, _s;
+	double _a, _b, _s;
 	int _h, _w; // size of the view
 	int center_x, center_y; // not in use yet
 	COLORREF _c1, _c2;
